+++ pkg/providers/iam/testdata/TestResourceAPIClient/importable_no_ipacl_actions_purgeoptions.tf
@@ -1,0 +1,41 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_iam_api_client" "test" {
+  authorized_users    = ["mw+2"]
+  client_type         = "CLIENT"
+  client_name         = "mw+2_1"
+  notification_emails = ["mw+2@example.com"]
+  client_description  = "Test API Client"
+  lock                = false
+  credential          = {}
+  group_access = {
+    clone_authorized_user_groups = false
+    groups = [
+      {
+        group_id = 123
+        role_id  = 340
+        sub_groups = [
+          {
+            group_id = 333
+            role_id  = 540
+          }
+        ]
+      }
+    ]
+  }
+  api_access = {
+    all_accessible_apis = false
+    apis = [
+      {
+        api_id       = 5580
+        access_level = "READ-ONLY"
+      },
+      {
+        api_id       = 5801
+        access_level = "READ-WRITE"
+      }
+    ]
+  }
+}